// src/app/model/broadcastMessage.ts
export class BroadCastMessage  
{  
        itbId  : number;
        targetAudience  : string;
        targetAudienceBoolean  : boolean;
        deptId  : string = null;
        subject  : string;
        message  : string = '';
        broadcastType  : string;
        userId  : string;
        dateCreated  : string;
        status  : string;
        deptName  : string;
        startDate: string;
        endDate: string;
        notifyTimeInterval: number;
        createdBy: string;
        starthour: string;
        startminute: string;
        startamOrPm: string;
        endhour: string = '';
        endminute: string = '';
        endamOrPm: string;
        readBroadCastMessage: boolean
}
        