// src/app/model/userDetails.ts
export class UserDetails {


        userName: string;
        roleId: null;
        userId: null;
        fullName: null;
        loginId: null;
        deptId: null;
        branchNo: null;
        lastLoginTime: null;
        bankingDate: null;
        roleName: null;
        branchName: null;
        cbsStatus: string = '';
}